Create generic directory traversal method, introduce `buildboxcommon_direntwrapper`, and add support for `chmod`'ing directories.

See merge request BuildGrid/buildbox/buildbox-common!160

+++ buildboxcommon/buildboxcommon_client.h
@@ -69,7 +69,8 @@ class Client {
            std::shared_ptr<LocalContentAddressableStorage::StubInterface>
                localCasClient,
            std::shared_ptr<Capabilities::StubInterface> capabilitiesClient,
-           int64_t maxBatchTotalSizeBytes = s_bytestreamChunkSizeBytes)
+           int64_t maxBatchTotalSizeBytes =
+               static_cast<int64_t>(s_bytestreamChunkSizeBytes))
         : d_bytestreamClient(bytestreamClient), d_casClient(casClient),
           d_localCasClient(localCasClient),
           d_capabilitiesClient(capabilitiesClient),
